moved security config files to config package
