+++ src/main/java/echoboard/echoboard/security/config/ApplicationNoSecurity.java
@@ -1,4 +1,4 @@
-package echoboard.echoboard.security;
+package echoboard.echoboard.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
