+++ src/main/java/echoboard/echoboard/security/config/DevelopmentSecurityConfiguration.java
@@ -1,4 +1,4 @@
-package echoboard.echoboard.security;
+package echoboard.echoboard.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
