+++ src/main/java/echoboard/echoboard/security/config/SecurityConfiguration.java
@@ -1,5 +1,6 @@
-package echoboard.echoboard.security;
+package echoboard.echoboard.security.config;
 
+import echoboard.echoboard.security.CustomAuthenticationSuccessHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
